packages/observability/pyproject.toml
[project]
name = "oie-observability"
version = "0.1.0"
description = "OIE observability stack with OpenTelemetry and structured logging"
requires-python = ">=3.12"
license = { text = "Proprietary" }
dependencies = [
    "opentelemetry-api",
    "opentelemetry-sdk",
    "opentelemetry-instrumentation-fastapi",
    "opentelemetry-exporter-otlp",
    "structlog",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff",
]

[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
